Only have one button to enable/disable all applications
