+++ src/app/simulation/applications/AvailableApplicationList.scss
@@ -1,10 +1,22 @@
-.applications {
-    border-collapse: collapse;
-    width: 100%;
+.available-application-list {
+    position: relative;
+    display: flex;
+    align-items: center;
+    border-bottom: 2px solid #CECECE;
+
+    .available-application-list__action {
+        width: 50px;
+    }
+
+    table {
+        border-collapse: collapse;
+        width: calc(100% - 50px);
+    }
 
     thead {
         background-color: #3f51b5;
         color: #fff;
+        border-left: 2px solid #CECECE;
     }
 
     th,
